add get data method
