+++ backend/src/libs/model/event.model.ts
@@ -1,5 +1,7 @@
 import { HttpError } from '@libs/api-gateway';
-import IParams from './params';
+import { tableNames } from '@libs/tableNames';
+import { stringifiedJson } from 'aws-sdk/clients/customerprofiles';
+import IParams, { IFetchEventIdParams, IFetchEventParams } from './params';
 
 const AWS = require('aws-sdk');
 
@@ -17,7 +19,10 @@ class EventModel {
     return await this.dynamodb.get(params).promise();
   }
 
-  public async getDataWithQuery(params: IParams, notFoundErrorMessage: string) {
+  private async getDataWithQuery(
+    params: IParams,
+    notFoundErrorMessage: string
+  ) {
     const data = await this.query(params);
     if (data.Items.length === 0) {
       throw new HttpError(404, notFoundErrorMessage);
@@ -25,12 +30,43 @@ class EventModel {
     return data;
   }
 
-  public async getDataWithGet(params: IParams, notFoundErrorMessage: string) {
+  private async getDataWithGet(params: IParams, notFoundErrorMessage: string) {
     const data = await this.get(params);
     if (Object.keys(data).length === 0) {
       throw new HttpError(404, notFoundErrorMessage);
     }
     return data;
+  }
+
+  public async getEventIdData(
+    userId: stringifiedJson,
+    notFoundErrorMessage: string
+  ) {
+    const fetchEventIdParams: IFetchEventIdParams = {
+      TableName: tableNames.USER_EVENT,
+      KeyConditionExpression: 'PK = :PK',
+      ExpressionAttributeValues: {
+        ':PK': userId,
+      },
+    };
+
+    return await this.getDataWithQuery(
+      fetchEventIdParams,
+      notFoundErrorMessage
+    );
+  }
+
+  public async getEventData(eventId: string, notFoundErrorMessage: string) {
+    const fetchEventParams: IFetchEventParams = {
+      TableName: tableNames.USER_EVENT,
+      Key: {
+        PK: 'event',
+        SK: eventId,
+      },
+      ProjectionExpression: `startingTimeReception,startingTimeWedding,message,address,dateWeddingReception,endingTimeReception,isEditable,groom,bride,dateWedding,endingTimeWedding,SK`,
+    };
+
+    return await this.getDataWithGet(fetchEventParams, notFoundErrorMessage);
   }
 }
